refactor: move autocrafter code to separate package
