+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/autocrafter/LockMode.java
@@ -1,4 +1,4 @@
-package com.refinedmods.refinedstorage.common.autocrafting;
+package com.refinedmods.refinedstorage.common.autocrafting.autocrafter;
 
 public enum LockMode {
     NEVER,
