+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/content/Blocks.java
@@ -1,6 +1,6 @@
 package com.refinedmods.refinedstorage.common.content;
 
-import com.refinedmods.refinedstorage.common.autocrafting.AutocrafterBlock;
+import com.refinedmods.refinedstorage.common.autocrafting.autocrafter.AutocrafterBlock;
 import com.refinedmods.refinedstorage.common.autocrafting.PatternGridBlock;
 import com.refinedmods.refinedstorage.common.constructordestructor.AbstractConstructorBlockEntity;
 import com.refinedmods.refinedstorage.common.constructordestructor.AbstractDestructorBlockEntity;
